+++ Backend/Controllers/wordsList.controller.ts
@@ -15,6 +15,24 @@ export class WordsListController {
     add: RequestHandler = async (req, res, next) => {
         try {
             const { name, toLanguage, fromLanguage } = req.body;
+            if(!toLanguage){
+                throw new ApiException("No target language provided", 400)
+            }
+            if(toLanguage.length > 6){
+                throw new ApiException("Length of target language is too big", 400)
+            }
+            if(!fromLanguage){
+                throw new ApiException("No source language provided", 400)
+            }
+            if(fromLanguage.length > 6){
+                throw new ApiException("Length of source language is too big", 400)
+            }
+            if(!name){
+                throw new ApiException("No word list name provided", 400)
+            }
+            if(name.length > 99){
+                throw new ApiException("Word list name is too long", 400)
+            }
             const createdList = await this.wordsListService.add(name, toLanguage, fromLanguage);
             return res.status(201).json({
                 success: true,
@@ -34,6 +52,15 @@ export class WordsListController {
     rename: RequestHandler = async (req, res, next) => {
         try {
             const { id, newName } = req.body;
+            if(!id){
+                throw new ApiException("No word list ID provided", 400)
+            }
+            if(!newName){
+                throw new ApiException("No new word list name provided", 400)
+            }
+            if(newName > 99){
+                throw new ApiException("New word list name is too long", 400)
+            }
             const renamedList = await this.wordsListService.rename(id, newName)
             return res.status(200).json({
                 success: true,
@@ -53,6 +80,9 @@ export class WordsListController {
     remove: RequestHandler = async (req, res, next) => {
         try {
             const { id } = req.body;
+            if(!id){
+                throw new ApiException("No word list ID provided", 400)
+            }
             const removedList = await this.wordsListService.remove(id);
             return res.status(201).json({
                 success: true,
@@ -90,6 +120,9 @@ export class WordsListController {
     createQuiz: RequestHandler = async (req, res, next) => {
         try {
             const { listId, quizType } = req.query
+            if(!listId){
+                throw new ApiException("No word list ID provided", 400)
+            }
             let quiz:QuizQuestion[]
             if (Number(quizType) == 1)
                 quiz = await this.wordsListService.createQuizOneKnown(listId)
